The central data structure of the MultiAmplicon package

The MultiAmplicon class is a container that stores at least primer pairs, read files and progressively processed data in an 'amplicon x samples' format. The slots in this object are incrementally filled with by running wrappers functions (mostly around functions from the dada2 package).

Subsetting for MultiAmplicon objects should conveniently subset all (potentially) filled slots

MultiAmplicon(PrimerPairsSet = PrimerPairsSet(),
  PairedReadFileSet = PairedReadFileSet(), rawCounts = matrix(),
  stratifiedFiles = list(), derep = list(), dada = list(),
  mergers = list(), sequenceTable = list(), sequenceTableNoChime = list(),
  sequenceTableFilled = list())

# S4 method for MultiAmplicon
colnames(x)

# S4 method for MultiAmplicon
rownames(x)

# S4 method for MultiAmplicon
ncol(x)

# S4 method for MultiAmplicon
nrow(x)

# S4 method for MultiAmplicon
dim(x)

getRawCounts(MA)

getStratifiedFilesF(MA, simplify = TRUE)

getStratifiedFilesR(MA, simplify = TRUE)

getDerepF(MA, simplify = TRUE)

getDerepR(MA, simplify = TRUE)

getDadaF(MA, simplify = TRUE)

getDadaR(MA, simplify = TRUE)

getMergers(MA, simplify = TRUE)

getSequenceTable(MA, simplify = TRUE)

getSequenceTableNoChime(MA, simplify = TRUE)

getSequencesFromTable(MA)

# S4 method for MultiAmplicon
getSequencesFromTable(MA)

# S4 method for MultiAmplicon
show(object)

# S4 method for MultiAmplicon,index,index,ANY
[(x, i, j, ..., drop = FALSE)

# S4 method for MultiAmplicon,index,missing,ANY
[(x, i, j, ..., drop = FALSE)

# S4 method for MultiAmplicon,missing,index,ANY
[(x, i, j, ..., drop = FALSE)

# S4 method for MultiAmplicon,missing,missing,ANY
[(x, i, j, ..., drop = FALSE)

Arguments

PrimerPairsSet
a set of primer pairs specifiying your amplicons see PrimerPairsSet-class

PairedReadFileSet
a set of paired end sequencing data files PairedReadFileSet-class

rawCounts
Users should not supply this parameter, the slot is created by sortAmplicons.

stratifiedFiles
Users should not supply this parameter, the slot is created by sortAmplicons.

derep
Users should not supply this parameter, the slot is created by derepMulti

dada
Users should not supply this parameter, the slot is created by dadaMulti

mergers
Users should not supply this parameter, the slot is created by mergeMulti

sequenceTable
Users should not supply this parameter, the slot is created by makeSequenceTableMulti

sequenceTableNoChime
Users should not supply this parameter, the slot is created by removeChimeraMulti

sequenceTableFilled
Users should not supply this parameter, the slot is created by fillSampleTables

x
MultiAmplicon-class object

MA
MultiAmplicon-class object

simplify
Should a list of objects be simplified to only one object if it has length one?

object
A MultiAmplicon-class object.

i
numeric, logical or names vector for subsetting rows (== amplicons)

j
numeric, logical or names vector for subsetting columns (== read files, corresponding usually to samples)

...
not used

drop
should not be used

Functions

MultiAmplicon: Constructor for MultiAmplicon-class

Slots

PrimerPairsSet

The primer pairs used in your experiment to specify amplicons stored in a PrimerPairsSet-class object.

PairedReadFileSet

The (quality filtered) fastq files (one file pair for each sample) that store your sequencing data.

rawCounts

A numeric matrix of sequencing read counts per amplicon and sample. Created by the function sortAmplicons

stratifiedFiles

temporary files as a result of stratifying into amplicons and samples using the function sortAmplicons. Forward (sometimes called R1) and reverse (sometimes called R2) files are stored as a (amplicons x samples) matrix of PairedReadFileSet-class objects.

derep

A PairedDerep-class object containing pairs of derep-class objects created by dada2’s derepFastq function

dada

A PairedDada-class object containing pairs of dada-class objects created by dada2’s dada function

mergers

sequenceTable

sequenceTableNoChime

sequenceTableFilled

MultiAmplicon(PrimerPairsSet, PairedReadFileSet)

See also

Examples

primerF <- c("AGAGTTTGATCCTGGCTCAG", "ACTCCTACGGGAGGCAGC",
            "GAATTGACGGAAGGGCACC", "YGGTGRTGCATGGCCGYT")
primerR <- c("CTGCWGCCNCCCGTAGG", "GACTACHVGGGTATCTAATCC",
             "AAGGGCATCACAGACCTGTTAT", "TCCTTCTGCAGGTTCACCTAC")

PPS <- PrimerPairsSet(primerF, primerR)

fastq.dir <- system.file("extdata", "fastq", package = "MultiAmplicon")
fastq.files <- list.files(fastq.dir, full.names=TRUE)
Ffastq.file <- fastq.files[grepl("F_filt", fastq.files)]
Rfastq.file <- fastq.files[grepl("R_filt", fastq.files)]

PRF <- PairedReadFileSet(Ffastq.file, Rfastq.file)

MA <- MultiAmplicon(PPS, PRF)

## the number of amplicons (primer pairs)
nrow(MA)

#> [1] 4

## the number of samples (sequencing read file pairs)
ncol(MA)

#> [1] 7

## sort into amplicons
MA1 <- sortAmplicons(MA, filedir=tempfile(pattern = "dir"))

#> creating directory /tmp/Rtmpc4G5wm/dir8e17637c5a9b

#> finished sorting 0 sequencing reads for S00_F_filt.fastq.gz in 
#>  /localstorage/ele/git_projects/MultiAmplicon/inst/extdata/fastq/S00_F_filt.fastq.gz and 
#>   /localstorage/ele/git_projects/MultiAmplicon/inst/extdata/fastq/S00_R_filt.fastq.gz 
#>  into 0  amplicons and written into /tmp/Rtmpc4G5wm/dir8e17637c5a9b

#> finished sorting 0 sequencing reads for S01_F_filt.fastq.gz in 
#>  /localstorage/ele/git_projects/MultiAmplicon/inst/extdata/fastq/S01_F_filt.fastq.gz and 
#>   /localstorage/ele/git_projects/MultiAmplicon/inst/extdata/fastq/S01_R_filt.fastq.gz 
#>  into 0  amplicons and written into /tmp/Rtmpc4G5wm/dir8e17637c5a9b

#> finished sorting 9348 sequencing reads for S02_F_filt.fastq.gz in 
#>  /localstorage/ele/git_projects/MultiAmplicon/inst/extdata/fastq/S02_F_filt.fastq.gz and 
#>   /localstorage/ele/git_projects/MultiAmplicon/inst/extdata/fastq/S02_R_filt.fastq.gz 
#>  into 4  amplicons and written into /tmp/Rtmpc4G5wm/dir8e17637c5a9b

#> finished sorting 5 sequencing reads for S03_F_filt.fastq.gz in 
#>  /localstorage/ele/git_projects/MultiAmplicon/inst/extdata/fastq/S03_F_filt.fastq.gz and 
#>   /localstorage/ele/git_projects/MultiAmplicon/inst/extdata/fastq/S03_R_filt.fastq.gz 
#>  into 1  amplicons and written into /tmp/Rtmpc4G5wm/dir8e17637c5a9b

#> finished sorting 14 sequencing reads for S04_F_filt.fastq.gz in 
#>  /localstorage/ele/git_projects/MultiAmplicon/inst/extdata/fastq/S04_F_filt.fastq.gz and 
#>   /localstorage/ele/git_projects/MultiAmplicon/inst/extdata/fastq/S04_R_filt.fastq.gz 
#>  into 3  amplicons and written into /tmp/Rtmpc4G5wm/dir8e17637c5a9b

#> finished sorting 51777 sequencing reads for S05_F_filt.fastq.gz in 
#>  /localstorage/ele/git_projects/MultiAmplicon/inst/extdata/fastq/S05_F_filt.fastq.gz and 
#>   /localstorage/ele/git_projects/MultiAmplicon/inst/extdata/fastq/S05_R_filt.fastq.gz 
#>  into 3  amplicons and written into /tmp/Rtmpc4G5wm/dir8e17637c5a9b

#> finished sorting 17358 sequencing reads for S06_F_filt.fastq.gz in 
#>  /localstorage/ele/git_projects/MultiAmplicon/inst/extdata/fastq/S06_F_filt.fastq.gz and 
#>   /localstorage/ele/git_projects/MultiAmplicon/inst/extdata/fastq/S06_R_filt.fastq.gz 
#>  into 4  amplicons and written into /tmp/Rtmpc4G5wm/dir8e17637c5a9b

MA2 <- derepMulti(MA1)

#> calling derepFastq with default parameters

#> amplicon AGAGTTTGATCCTGGCTCAG.CTGCWGCCNCCCGTAGG dereplicating for 5 of 7 possible sample files.

#> calling derepFastq with default parameters

#> amplicon ACTCCTACGGGAGGCAGC.GACTACHVGGGTATCTAATCC dereplicating for 4 of 7 possible sample files.

#> calling derepFastq with default parameters

#> amplicon GAATTGACGGAAGGGCACC.AAGGGCATCACAGACCTGTTAT dereplicating for 2 of 7 possible sample files.

#> calling derepFastq with default parameters

#> amplicon YGGTGRTGCATGGCCGYT.TCCTTCTGCAGGTTCACCTAC dereplicating for 4 of 7 possible sample files.

MA3 <- dadaMulti(MA2, selfConsist = TRUE)

#> 
#> 
#> amplicon AGAGTTTGATCCTGGCTCAG.CTGCWGCCNCCCGTAGG: dada estimation of sequence variants from 5 of 7 possible sample files

#> calling dada with selfConsist=TRUE parameters

#> Initializing error rates to maximum possible estimate.
#> selfConsist step 1 .....
#>    selfConsist step 2
#>    selfConsist step 3
#>    selfConsist step 4
#>    selfConsist step 5
#>    selfConsist step 6
#> Convergence after  6  rounds.
#> Initializing error rates to maximum possible estimate.
#> selfConsist step 1 .....
#>    selfConsist step 2
#>    selfConsist step 3
#>    selfConsist step 4
#>    selfConsist step 5
#>    selfConsist step 6
#> Convergence after  6  rounds.

#> 
#> 
#> amplicon ACTCCTACGGGAGGCAGC.GACTACHVGGGTATCTAATCC: dada estimation of sequence variants from 4 of 7 possible sample files

#> calling dada with selfConsist=TRUE parameters

#> Initializing error rates to maximum possible estimate.
#> selfConsist step 1 ....
#>    selfConsist step 2
#>    selfConsist step 3
#>    selfConsist step 4
#>    selfConsist step 5
#> Convergence after  5  rounds.
#> Initializing error rates to maximum possible estimate.
#> selfConsist step 1 ....
#>    selfConsist step 2
#>    selfConsist step 3
#>    selfConsist step 4
#>    selfConsist step 5
#>    selfConsist step 6
#> Convergence after  6  rounds.

#> 
#> 
#> amplicon GAATTGACGGAAGGGCACC.AAGGGCATCACAGACCTGTTAT: dada estimation of sequence variants from 2 of 7 possible sample files

#> calling dada with selfConsist=TRUE parameters

#> Initializing error rates to maximum possible estimate.
#> selfConsist step 1 ..
#>    selfConsist step 2
#>    selfConsist step 3
#>    selfConsist step 4
#>    selfConsist step 5
#> Convergence after  5  rounds.
#> Initializing error rates to maximum possible estimate.
#> selfConsist step 1 ..
#>    selfConsist step 2
#>    selfConsist step 3
#>    selfConsist step 4
#> Convergence after  4  rounds.

#> 
#> 
#> amplicon YGGTGRTGCATGGCCGYT.TCCTTCTGCAGGTTCACCTAC: dada estimation of sequence variants from 4 of 7 possible sample files

#> calling dada with selfConsist=TRUE parameters

#> Initializing error rates to maximum possible estimate.
#> selfConsist step 1 ....
#>    selfConsist step 2
#>    selfConsist step 3
#>    selfConsist step 4
#>    selfConsist step 5
#>    selfConsist step 6
#> Convergence after  6  rounds.
#> Initializing error rates to maximum possible estimate.
#> selfConsist step 1 ....
#>    selfConsist step 2
#>    selfConsist step 3
#>    selfConsist step 4
#>    selfConsist step 5
#> Convergence after  5  rounds.

MA4 <- mergeMulti(MA3, justConcatenate=TRUE)

#> 
#> merging sequences from 5 samples for amplicon AGAGTTTGATCCTGGCTCAG.CTGCWGCCNCCCGTAGG

#> calling mergePairs with justConcatenate=TRUE parameters

#> 
#> merging sequences from 4 samples for amplicon ACTCCTACGGGAGGCAGC.GACTACHVGGGTATCTAATCC

#> calling mergePairs with justConcatenate=TRUE parameters

#> 
#> merging sequences from 2 samples for amplicon GAATTGACGGAAGGGCACC.AAGGGCATCACAGACCTGTTAT

#> calling mergePairs with justConcatenate=TRUE parameters

#> 
#> merging sequences from 4 samples for amplicon YGGTGRTGCATGGCCGYT.TCCTTCTGCAGGTTCACCTAC

#> calling mergePairs with justConcatenate=TRUE parameters

MA5 <- makeSequenceTableMulti(MA4)

#> calling makeSequenceTable with default parameters

#> calling makeSequenceTable with default parameters

#> calling makeSequenceTable with default parameters

#> calling makeSequenceTable with default parameters

MA6 <- removeChimeraMulti(MA5, mc.cores=1)

#> calling removeBimeraDenovo with default parameters

#> calling removeBimeraDenovo with default parameters

#> calling removeBimeraDenovo with default parameters

#> calling removeBimeraDenovo with default parameters

MA7 <- fillSampleTables(MA6)

#> 
#>  class Colnames: NULL 
#> 
#>  Colnames: 
#> 
#>  Which: