A class representing sequences of forward and reverse primers.

The PrimerPairsSet class is a container for storing primer pairs. This means exactly two DNAStringSet objects of the same length specifying primer-pairs. Primer sequences can be provided as character strings and will be converted to DNAStringSet by the constructor function of the same name. primerF and primerR have to be of the same length to specify primer pairs. Warnings are given if primer sequences are of unusal length (<16 or >26 bases).

Accessor like functions length gives the number of read paires in a PrimerPairsSet-class object names of primer-pairs (amplicons) in a PrimerPairsSet-class object

PrimerPairsSet(primerF, primerR)

# S4 method for PrimerPairsSet
length(x)

# S4 method for PrimerPairsSet
names(x)

# S4 method for PrimerPairsSet
show(object)

# S4 method for PrimerPairsSet,index,missing,ANY
[(x, i, j, ..., drop = NA)

Arguments

primerF
Character vector or DNAStringSet. Can be named or unnamed.

primerR
Character vector or DNAStringSet of the same length. Can be named or unnamed.

x
A PrimerPairsSet-class object.

object
A PrimerPairsSet-class object.

i
integer, or logical value which primer pairs to select or character string giving the name of the primer pair

j
not used

...
not used

drop
not used

Value

PrimerPairsSet-class

Slots

primerF

DNAStringSet. Can be named or unnamed.

primerR

DNAStringSet of the same length. Can be named or unnnamed.

names

Character string. Either constructed as a concatenation of names of forward and reverse primers or of their sequences (if primer sequences are unnamed).

.mapF

(automatically generated) maps potentially duplicate entries in FW primers to unique entries.

.mapR

(auto-generated) maps potentially duplicate entries in FW primers to unique entries.

.uniqueF

(auto-generated) unique forward primers as character strings.

.uniqueR

(auto-generated) unique reverse primers as character strings.

See also

DNAStringSet