mergeMulti

merge denoised pairs of forward and reverse reads inside an MultiAmplcion object.

mergeMulti(MA, ...)

Arguments

MA
MultiAmplicon-class object with derep and dada slots filled.

...
additional arguments to be passed to the mergePairs function of dada2, all arguments to the function can be given as a vector of the same length as the number of primer pairs in the MultiAmplicon object, allowing to specify e.g. justConcatenate to be set TRUE for only some of the amplicons, or to specify different minOverlap for each amplicon. If a shorter vector is given it will be recycled to match the number of ampicons.

Value

A MultiAmplicon-class object with the mergers slot filled

Details

This is a wrapper for mergePairs from dada2. It works on an MultiAmplicon-class object with derep and dada slots filled. Use dadaMulti and derepMulti on a amplicon sorted (see sortAmplicons) MultiAmplicon-class object to preprocess your multi-marker data to this point.