noChimeMulti

Remove chimeric sequencing read pairs from a MultiAmplicon object.

noChimeMulti(MA, mc.cores = getOption("mc.cores", 2L), ...)

Arguments

MA
A MultiAmplicon-class object preprocessed to have a sequenceTableMulti populated

mc.cores
integer number of cores to use for parallelization

...
passed on to removeBimeraDenovo. All arguments to the function can be given as a vector of the same length as the number of primer pairs in the MultiAmplicon object, allowing to specify different parameters for each amplicon. If a shorter vector is given it will be recycled to match the number of amplicons.

Value

a MultiAmplicon-class object with the sequenceTableNoChime filled

Details

This is a wrapper for the removeBimeraDenovo function of dada2. It works on an MultiAmplicon-class object with the sequenceTable slot filled. Use dadaMulti, derepMulti, mergeMulti and sequenceTableMulti on a amplicon sorted (see sortAmplicons) MultiAmplicon-class object to preprocess your multi-marker data to this point.