plotAmpliconNumbers

Plot the raw counts for primer pairs matched in different sample files as a heatmap

plotAmpliconNumbers(MA, transf = function(x) log10(x + 1), ...)

# S4 method for matrix
plotAmpliconNumbers(MA, transf = function(x) log10(x + 1),
  ...)

# S4 method for MultiAmplicon
plotAmpliconNumbers(MA, transf = function(x) log10(x
  + 1), ...)

Arguments

MA
The primer matching (amplicon) matrix of a MultiAmplicon object or the object itself

transf
transformation to be applied, recommended default is log10(x+1)

...
additional parameter to be passed to pheatmap function

Value

just like the original pheatmap function, A heatmap is drawn to the graphics output device. Returned are (invisibly) a list of components: 1."tree_row" the clustering of rows as hclust object 2. "tree_col" the clustering of columns as hclust object 3. "kmeans" the kmeans clustering of rows if parameter kmeans_k was specified.

Details

The function uses pheatmap to plot the primer matching statistics stored in the rawCounts slot of a MultiAmplicons-class object