A class combining sequences of forward and reverse primers (in a PrimerPairsSet-class) plus file names of paired end sequencing files (in a PairedReadFileSet-class).

The MultiAmplicon class is a container for storing primer pairs, read files and processed data in an 'amplicon x samples' format. The object is incrementally filled by running functions (from the package dada2) with additional slots.

Subsetting for MultiAmplicon objects should conveniently subset all (potentially) filled slots

MultiAmplicon(PrimerPairsSet = PrimerPairsSet(),
  PairedReadFileSet = PairedReadFileSet(), rawCounts = matrix(),
  stratifiedFiles = list(), derep = list(), dada = list(),
  mergers = list(), sequenceTable = list(), sequenceTableNoChime = list(),
  sequenceTableFilled = list())

# S4 method for MultiAmplicon
colnames(x)

# S4 method for MultiAmplicon
rownames(x)

# S4 method for MultiAmplicon
ncol(x)

# S4 method for MultiAmplicon
nrow(x)

# S4 method for MultiAmplicon
dim(x)

rawCounts(x)

# S4 method for MultiAmplicon
rawCounts(x)

getStratifiedFilesF(MA)

getStratifiedFilesR(MA)

getDerepF(MA)

getDerepR(MA)

getDadaF(MA)

getDadaR(MA)

getMergers(MA)

getSequenceTable(MA)

getSequenceTableNoChime(MA)

getSequencesFromTable(MA)

# S4 method for MultiAmplicon
getSequencesFromTable(MA)

# S4 method for MultiAmplicon
show(object)

# S4 method for MultiAmplicon,index,index,ANY
[(x, i, j, ..., drop = FALSE)

# S4 method for MultiAmplicon,index,missing,ANY
[(x, i, j, ..., drop = FALSE)

# S4 method for MultiAmplicon,missing,index,ANY
[(x, i, j, ..., drop = FALSE)

# S4 method for MultiAmplicon,missing,missing,ANY
[(x, i, j, ..., drop = FALSE)

Arguments

PrimerPairsSet
a set of primer pairs specifiying your amplicons see PrimerPairsSet-class

PairedReadFileSet
a set of paired end sequencing data files PairedReadFileSet-class

rawCounts
Users should not supply this parameter, the slot is created by sortAmplicons.

stratifiedFiles
Users should not supply this parameter, the slot is created by sortAmplicons.

derep
Users should not supply this parameter, the slot is created by derepMulti

dada
Users should not supply this parameter, the slot is created by dadaMulti

mergers
Users should not supply this parameter, the slot is created by mergeMulti

sequenceTable
Users should not supply this parameter, the slot is created by makeSequenceTableMulti

sequenceTableNoChime
Users should not supply this parameter, the slot is created by removeChimeraMulti

sequenceTableFilled
Users should not supply this parameter, the slot is created by fillSampleTables

x
MultiAmplicon-class object

MA
MultiAmplicon-class object

object
A MultiAmplicon-class object.

i
numeric, logical or names vector for subsetting rows (== amplicons)

j
numeric, logical or names vector for subsetting columns (== read files, corresponding usually to samples)

...
not used

drop
should not be used

Functions

MultiAmplicon: Constructor for MultiAmplicon-class

Slots

PrimerPairsSet

The primer pairs used in your experiment to specify amplicons stored in a PrimerPairsSet-class object.

PairedReadFileSet

The (quality filtered) fastq files (one file pair for each sample) that store your sequencing data.

rawCounts

A numeric matrix of sequencing read counts per amplicon and sample. Created by the function sortAmplicons

stratifiedFiles

temporary files as a result of stratifying into amplicons and samples using the function sortAmplicons. Forward (sometimes called R1) and reverse (sometimes called R2) files are stored as a (amplicons x samples) matrix of PairedReadFileSet-class objects.

derep

A PairedDerep-class object containing pairs of derep-class objects created by dada2’s derepFastq function

dada

A PairedDada-class object containing pairs of dada-class objects created by dada2’s dada function

mergers

sequenceTable

sequenceTableNoChime

sequenceTableFilled

MultiAmplicon(PrimerPairsSet, PairedReadFileSet)

See also

derepFastq,dada

Examples

primerF <- c("AGAGTTTGATCCTGGCTCAG", "ACTCCTACGGGAGGCAGC",
            "GAATTGACGGAAGGGCACC", "YGGTGRTGCATGGCCGYT")
primerR <- c("CTGCWGCCNCCCGTAGG", "GACTACHVGGGTATCTAATCC",
             "AAGGGCATCACAGACCTGTTAT", "TCCTTCTGCAGGTTCACCTAC")

PPS <- PrimerPairsSet(primerF, primerR)

fastq.dir <- system.file("extdata", "fastq", package = "MultiAmplicon")
fastq.files <- list.files(fastq.dir, full.names=TRUE)
Ffastq.file <- fastq.files[grepl("F_filt", fastq.files)]
Rfastq.file <- fastq.files[grepl("R_filt", fastq.files)]

PRF <- PairedReadFileSet(Ffastq.file, Rfastq.file)

MA <- MultiAmplicon(PPS, PRF)

## the number of amplicons (primer pairs)
nrow(MA)

#> [1] 4

## the number of samples (sequencing read file pairs)
ncol(MA)

#> [1] 7

## sort into amplicons
MA1 <- sortAmplicons(MA)

#> using existing directory /tmp/RtmpeEHDm3 and adding specific prefixes to avoid problems when running code twice

#> finished sorting 0 sequencing reads for S00_F_filt.fastq.gz in
#>  /localstorage/ele/git_projects/MultiAmplicon/inst/extdata/fastq/S00_F_filt.fastq.gz and 
#>  /localstorage/ele/git_projects/MultiAmplicon/inst/extdata/fastq/S00_R_filt.fastq.gz
#>  into 0 amplicons

#> finished sorting 0 sequencing reads for S01_F_filt.fastq.gz in
#>  /localstorage/ele/git_projects/MultiAmplicon/inst/extdata/fastq/S01_F_filt.fastq.gz and 
#>  /localstorage/ele/git_projects/MultiAmplicon/inst/extdata/fastq/S01_R_filt.fastq.gz
#>  into 0 amplicons

#> finished sorting 9348 sequencing reads for S02_F_filt.fastq.gz in
#>  /localstorage/ele/git_projects/MultiAmplicon/inst/extdata/fastq/S02_F_filt.fastq.gz and 
#>  /localstorage/ele/git_projects/MultiAmplicon/inst/extdata/fastq/S02_R_filt.fastq.gz
#>  into 4 amplicons

#> finished sorting 5 sequencing reads for S03_F_filt.fastq.gz in
#>  /localstorage/ele/git_projects/MultiAmplicon/inst/extdata/fastq/S03_F_filt.fastq.gz and 
#>  /localstorage/ele/git_projects/MultiAmplicon/inst/extdata/fastq/S03_R_filt.fastq.gz
#>  into 1 amplicons

#> finished sorting 14 sequencing reads for S04_F_filt.fastq.gz in
#>  /localstorage/ele/git_projects/MultiAmplicon/inst/extdata/fastq/S04_F_filt.fastq.gz and 
#>  /localstorage/ele/git_projects/MultiAmplicon/inst/extdata/fastq/S04_R_filt.fastq.gz
#>  into 3 amplicons

#> finished sorting 51777 sequencing reads for S05_F_filt.fastq.gz in
#>  /localstorage/ele/git_projects/MultiAmplicon/inst/extdata/fastq/S05_F_filt.fastq.gz and 
#>  /localstorage/ele/git_projects/MultiAmplicon/inst/extdata/fastq/S05_R_filt.fastq.gz
#>  into 3 amplicons

#> finished sorting 17358 sequencing reads for S06_F_filt.fastq.gz in
#>  /localstorage/ele/git_projects/MultiAmplicon/inst/extdata/fastq/S06_F_filt.fastq.gz and 
#>  /localstorage/ele/git_projects/MultiAmplicon/inst/extdata/fastq/S06_R_filt.fastq.gz
#>  into 4 amplicons

MA2 <- derepMulti(MA1)

MA3 <- dadaMulti(MA2, selfConsist = TRUE)

#> 
#> 
#> amplicon AGAGTTTGATCCTGGCTCAG.CTGCWGCCNCCCGTAGG: dada estimation of sequence variants from 5 of 7 possible sample files

#> calling dada with selfConsist=TRUE parameters

#> Initializing error rates to maximum possible estimate.
#> Sample 1 - 3495 reads in 1493 unique sequences.
#> Sample 2 - 5 reads in 5 unique sequences.
#> Sample 3 - 5 reads in 5 unique sequences.
#> Sample 4 - 29310 reads in 7224 unique sequences.
#> Sample 5 - 2569 reads in 1053 unique sequences.
#>    selfConsist step 2 
#>    selfConsist step 3 
#>    selfConsist step 4 
#>    selfConsist step 5 
#> Convergence after  5  rounds.
#> Initializing error rates to maximum possible estimate.
#> Sample 1 - 3495 reads in 1535 unique sequences.
#> Sample 2 - 5 reads in 5 unique sequences.
#> Sample 3 - 5 reads in 4 unique sequences.
#> Sample 4 - 29310 reads in 8273 unique sequences.
#> Sample 5 - 2569 reads in 994 unique sequences.
#>    selfConsist step 2 
#>    selfConsist step 3 
#>    selfConsist step 4 
#>    selfConsist step 5 
#> Convergence after  5  rounds.

#> 
#> 
#> amplicon ACTCCTACGGGAGGCAGC.GACTACHVGGGTATCTAATCC: dada estimation of sequence variants from 4 of 7 possible sample files

#> calling dada with selfConsist=TRUE parameters

#> Initializing error rates to maximum possible estimate.
#> Sample 1 - 2949 reads in 1142 unique sequences.
#> Sample 2 - 2 reads in 2 unique sequences.
#> Sample 3 - 20391 reads in 4136 unique sequences.
#> Sample 4 - 2690 reads in 968 unique sequences.
#>    selfConsist step 2 
#>    selfConsist step 3 
#>    selfConsist step 4 
#> Convergence after  4  rounds.
#> Initializing error rates to maximum possible estimate.
#> Sample 1 - 2949 reads in 1290 unique sequences.
#> Sample 2 - 2 reads in 2 unique sequences.
#> Sample 3 - 20391 reads in 5719 unique sequences.
#> Sample 4 - 2690 reads in 1100 unique sequences.
#>    selfConsist step 2 
#>    selfConsist step 3 
#>    selfConsist step 4 
#>    selfConsist step 5 
#> Convergence after  5  rounds.

#> 
#> 
#> amplicon GAATTGACGGAAGGGCACC.AAGGGCATCACAGACCTGTTAT: dada estimation of sequence variants from 2 of 7 possible sample files

#> calling dada with selfConsist=TRUE parameters

#> Initializing error rates to maximum possible estimate.
#> Sample 1 - 1817 reads in 589 unique sequences.
#> Sample 2 - 11671 reads in 3783 unique sequences.
#>    selfConsist step 2 
#>    selfConsist step 3 
#>    selfConsist step 4 
#> Convergence after  4  rounds.
#> Initializing error rates to maximum possible estimate.
#> Sample 1 - 1817 reads in 547 unique sequences.
#> Sample 2 - 11671 reads in 4054 unique sequences.
#>    selfConsist step 2 
#>    selfConsist step 3 
#>    selfConsist step 4 
#> Convergence after  4  rounds.

#> 
#> 
#> amplicon YGGTGRTGCATGGCCGYT.TCCTTCTGCAGGTTCACCTAC: dada estimation of sequence variants from 4 of 7 possible sample files

#> calling dada with selfConsist=TRUE parameters

#> Initializing error rates to maximum possible estimate.
#> Sample 1 - 1087 reads in 394 unique sequences.
#> Sample 2 - 7 reads in 6 unique sequences.
#> Sample 3 - 2076 reads in 959 unique sequences.
#> Sample 4 - 428 reads in 166 unique sequences.
#>    selfConsist step 2 
#>    selfConsist step 3 
#>    selfConsist step 4 
#> Convergence after  4  rounds.
#> Initializing error rates to maximum possible estimate.
#> Sample 1 - 1087 reads in 433 unique sequences.
#> Sample 2 - 7 reads in 5 unique sequences.
#> Sample 3 - 2076 reads in 962 unique sequences.
#> Sample 4 - 428 reads in 183 unique sequences.
#>    selfConsist step 2 
#>    selfConsist step 3 
#>    selfConsist step 4 
#> Convergence after  4  rounds.

MA4 <- mergeMulti(MA3, justConcatenate=TRUE)

#> 
#> merging sequences from 5 samples for amplicon AGAGTTTGATCCTGGCTCAG.CTGCWGCCNCCCGTAGG

#> calling mergePairs with justConcatenate=TRUE parameters

#> 
#> merging sequences from 4 samples for amplicon ACTCCTACGGGAGGCAGC.GACTACHVGGGTATCTAATCC

#> calling mergePairs with justConcatenate=TRUE parameters

#> 
#> merging sequences from 2 samples for amplicon GAATTGACGGAAGGGCACC.AAGGGCATCACAGACCTGTTAT

#> calling mergePairs with justConcatenate=TRUE parameters

#> 
#> merging sequences from 4 samples for amplicon YGGTGRTGCATGGCCGYT.TCCTTCTGCAGGTTCACCTAC

#> calling mergePairs with justConcatenate=TRUE parameters

MA5 <- makeSequenceTableMulti(MA4)

#> calling makeSequenceTable with default parameters

#> calling makeSequenceTable with default parameters

#> calling makeSequenceTable with default parameters

#> calling makeSequenceTable with default parameters

MA6 <- removeChimeraMulti(MA5, mc.cores=1)

#> calling removeBimeraDenovo with default parameters

#> As of the 1.4 release, the default method changed to consensus (from pooled).

#> calling removeBimeraDenovo with default parameters

#> As of the 1.4 release, the default method changed to consensus (from pooled).

#> calling removeBimeraDenovo with default parameters

#> As of the 1.4 release, the default method changed to consensus (from pooled).

#> calling removeBimeraDenovo with default parameters

#> As of the 1.4 release, the default method changed to consensus (from pooled).

MA7 <- fillSampleTables(MA6)