A class representing file paths to paired end reads.

Two character vectors of the same length specifying file names of paired end reads can be stored in this class. Filenames a checked for their existence. Usually these are the filnames of quality filtered fastq files already stratified into samples (one file pair for each sample).

PairedReadFileSet(readsF = character(), readsR = character())

# S4 method for PairedReadFileSet
length(x)

# S4 method for PairedReadFileSet
show(object)

# S4 method for PairedReadFileSet,index,missing,ANY
[(x, i, j, ..., drop = TRUE)

Arguments

readsF
The path and filenames containing forward (R1) reads

readsR
The path and filenames containing reverse (R2) reads

x
PairedReadFileSet-class object

object
A PairedReadFileSet-class object.

i
numeric to select

j
not used

...
not used

drop
not used

Value

PairedReadFileSet

Functions

PairedReadFileSet: Constructor for PairedReadFileSet-class

Slots

readF

A character vector specifying the file paths to files containing forward (sometimes called R1) sequencing reads. This vector can be named to store short short-name of samples.

readR

A character vector specifiying the file path to a file containing reverse (sometimes called R2) sequencing reads. This vector can be named to store short short-name of samples. Names of readF and readR should be identical in this case

names

Is either (first choice) from names of the forward reads, or (if these are empty) constructed from basename of forward read files (filename without directory).

Author

Emanuel Heitlinger