sortAmplicons

Sort different amplicons into a fully stratified samples x amplicons structure based on primer matches.

sortAmplicons(MA, n = 1e+06, countOnly = FALSE, rmPrimer = TRUE,
  filedir = tempdir(), ...)

# S4 method for MultiAmplicon
sortAmplicons(MA, n = 1e+06, countOnly = FALSE,
  rmPrimer = TRUE, filedir = tempdir(), ...)

Arguments

MA
MultiAmplicon-class object containing a set of paired end files and a primer-pairs set.

n
parameter passed to the yield functions of package ShortRead. This controls the memory consumption during streaming. Lower values result in lower memory requirements but might result longer processing time due to more repeated I/O operations reading the sequence files.

countOnly
logical argument if set TRUE only a matrix of read counts is returned

rmPrimer
logical, indicating whether primer sequences should be removed during sorting

filedir
path to an existing or newly to be created folder on your computer. tempfile is used within this folder to create unique filnames trying to avoid problems in case the folder has been used before.

...
additional parameter so be passed to Biostrings::isMatchingStartingAt. Be careful when using multiple starting positions or allowing error. This could lead to read pairs being assigned to multiple amplicons.

Value

MultiAmplicon: By default (countOnly=FALSE) a MultiAmplicon-class object is returned with the stratifiedFiles slot populated. Stratified file names are constructed using a unique string created by tempfile and stored in the given filedir (by default R's tempdir). If the countOnly is set only a numeric matrix of read counts is returned.

Details

This function uses isMatchingStartingAt to match primer sequences at the first position of forward and reverse sequences. These primer sequences can be removed. The remaining sequences of interest are written to (temporary) files to allow processing via standard metabarcoding pipelines.

Examples

primerF <- c("AGAGTTTGATCCTGGCTCAG", "ACTCCTACGGGAGGCAGC",
            "GAATTGACGGAAGGGCACC", "YGGTGRTGCATGGCCGYT")
primerR <- c("CTGCWGCCNCCCGTAGG", "GACTACHVGGGTATCTAATCC",
             "AAGGGCATCACAGACCTGTTAT", "TCCTTCTGCAGGTTCACCTAC")

PPS <- PrimerPairsSet(primerF, primerR)

fastq.dir <- system.file("extdata", "fastq", package = "MultiAmplicon")
fastq.files <- list.files(fastq.dir, full.names=TRUE)
Ffastq.file <- fastq.files[grepl("F_filt", fastq.files)]
Rfastq.file <- fastq.files[grepl("R_filt", fastq.files)]

PRF <- PairedReadFileSet(Ffastq.file, Rfastq.file)

MA <- MultiAmplicon(PPS, PRF)

## sort into amplicons
MA1 <- sortAmplicons(MA)

#> using existing directory /tmp/Rtmpt5VHLH and adding specific prefixes to avoid problems when running code twice

#> finished sorting 0 sequencing reads for S00_F_filt.fastq.gz in
#>  /localstorage/ele/git_projects/MultiAmplicon/inst/extdata/fastq/S00_F_filt.fastq.gz and 
#>  /localstorage/ele/git_projects/MultiAmplicon/inst/extdata/fastq/S00_R_filt.fastq.gz
#>  into 0 amplicons

#> finished sorting 0 sequencing reads for S01_F_filt.fastq.gz in
#>  /localstorage/ele/git_projects/MultiAmplicon/inst/extdata/fastq/S01_F_filt.fastq.gz and 
#>  /localstorage/ele/git_projects/MultiAmplicon/inst/extdata/fastq/S01_R_filt.fastq.gz
#>  into 0 amplicons

#> finished sorting 9348 sequencing reads for S02_F_filt.fastq.gz in
#>  /localstorage/ele/git_projects/MultiAmplicon/inst/extdata/fastq/S02_F_filt.fastq.gz and 
#>  /localstorage/ele/git_projects/MultiAmplicon/inst/extdata/fastq/S02_R_filt.fastq.gz
#>  into 4 amplicons

#> finished sorting 5 sequencing reads for S03_F_filt.fastq.gz in
#>  /localstorage/ele/git_projects/MultiAmplicon/inst/extdata/fastq/S03_F_filt.fastq.gz and 
#>  /localstorage/ele/git_projects/MultiAmplicon/inst/extdata/fastq/S03_R_filt.fastq.gz
#>  into 1 amplicons

#> finished sorting 14 sequencing reads for S04_F_filt.fastq.gz in
#>  /localstorage/ele/git_projects/MultiAmplicon/inst/extdata/fastq/S04_F_filt.fastq.gz and 
#>  /localstorage/ele/git_projects/MultiAmplicon/inst/extdata/fastq/S04_R_filt.fastq.gz
#>  into 3 amplicons

#> finished sorting 51777 sequencing reads for S05_F_filt.fastq.gz in
#>  /localstorage/ele/git_projects/MultiAmplicon/inst/extdata/fastq/S05_F_filt.fastq.gz and 
#>  /localstorage/ele/git_projects/MultiAmplicon/inst/extdata/fastq/S05_R_filt.fastq.gz
#>  into 3 amplicons

#> finished sorting 17358 sequencing reads for S06_F_filt.fastq.gz in
#>  /localstorage/ele/git_projects/MultiAmplicon/inst/extdata/fastq/S06_F_filt.fastq.gz and 
#>  /localstorage/ele/git_projects/MultiAmplicon/inst/extdata/fastq/S06_R_filt.fastq.gz
#>  into 4 amplicons